samples/issue-001/src/requirements.txt
azure-functions>=1.17.0
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
openai>=1.12.0
pypdf>=3.17.0
tiktoken>=0.5.0
